Moved FASTOADProblem to fastoad.openmdao
